backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
python-dotenv>=1.0.1
langchain>=0.3.20
langchain-openai>=0.3.0
langchain-google-genai>=2.0.8
langchain-community>=0.3.20
langgraph>=0.2.60
sse-starlette>=2.1.0
httpx>=0.27.0
PyJWT>=2.8.0
